+++ tests/Mock/ExclusionStrategies/FooDeserializationExclusionStrategy.php
@@ -8,15 +8,14 @@ namespace Tebru\Gson\Test\Mock\ExclusionStrategies;
 
 use Tebru\Gson\ClassMetadata;
 use Tebru\Gson\Exclusion\ClassDeserializationExclusionStrategy;
-use Tebru\Gson\Exclusion\ClassSerializationExclusionStrategy;
-use Tebru\Gson\Test\Mock\UserMock;
+use Tebru\Gson\Test\Mock\Foo;
 
 /**
- * Class UserMockExclusionStrategy
+ * Class FooDeserializationExclusionStrategy
  *
  * @author Nate Brunette <n@tebru.net>
  */
-class UserMockExclusionStrategy implements ClassSerializationExclusionStrategy, ClassDeserializationExclusionStrategy
+class FooDeserializationExclusionStrategy implements ClassDeserializationExclusionStrategy
 {
     /**
      * Returns true if the class should be skipped during deserialization
@@ -26,18 +25,7 @@ class UserMockExclusionStrategy implements ClassSerializationExclusionStrategy, 
      */
     public function skipDeserializingClass(ClassMetadata $class): bool
     {
-        return UserMock::class === $class->getName();
-    }
-
-    /**
-     * Returns true if the class should be skipped during serialization
-     *
-     * @param ClassMetadata $class
-     * @return bool
-     */
-    public function skipSerializingClass(ClassMetadata $class): bool
-    {
-        return UserMock::class === $class->getName();
+        return Foo::class === $class->getName();
     }
 
     /**