+++ tests/Mock/ExclusionStrategies/FooSerializationExclusionStrategy.php
@@ -7,28 +7,16 @@
 namespace Tebru\Gson\Test\Mock\ExclusionStrategies;
 
 use Tebru\Gson\ClassMetadata;
-use Tebru\Gson\Exclusion\ClassDeserializationExclusionStrategy;
 use Tebru\Gson\Exclusion\ClassSerializationExclusionStrategy;
-use Tebru\Gson\Test\Mock\UserMock;
+use Tebru\Gson\Test\Mock\Foo;
 
 /**
- * Class UserMockExclusionStrategy
+ * Class FooSerializationExclusionStrategy
  *
  * @author Nate Brunette <n@tebru.net>
  */
-class UserMockExclusionStrategy implements ClassSerializationExclusionStrategy, ClassDeserializationExclusionStrategy
+class FooSerializationExclusionStrategy implements ClassSerializationExclusionStrategy
 {
-    /**
-     * Returns true if the class should be skipped during deserialization
-     *
-     * @param ClassMetadata $class
-     * @return bool
-     */
-    public function skipDeserializingClass(ClassMetadata $class): bool
-    {
-        return UserMock::class === $class->getName();
-    }
-
     /**
      * Returns true if the class should be skipped during serialization
      *
@@ -37,7 +25,7 @@ class UserMockExclusionStrategy implements ClassSerializationExclusionStrategy, 
      */
     public function skipSerializingClass(ClassMetadata $class): bool
     {
-        return UserMock::class === $class->getName();
+        return Foo::class === $class->getName();
     }
 
     /**